+++ routes/web.php
@@ -3,8 +3,8 @@
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ListingController;
-use PhpParser\Node\Expr\List_;
 
 /*
 |--------------------------------------------------------------------------
@@ -32,22 +32,40 @@ use PhpParser\Node\Expr\List_;
 Route::get('/', [ListingController::class, 'index']);
 
 // show create listing form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // store new listing
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // show edit form
-Route::get('/listings/{id}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{id}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // update listing
-Route::put('/listings/{id}', [ListingController::class, 'update']);
+Route::put('/listings/{id}', [ListingController::class, 'update'])->middleware('auth');
 
 // delete listing
-Route::delete('/listings/{id}/delete', [ListingController::class, 'destroy']);
+Route::delete('/listings/{id}/delete', [ListingController::class, 'destroy'])->middleware('auth');
+
+// manage listing
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // single listing
 Route::get('/listings/{id}', [ListingController::class, 'show']);
+
+// show register form
+Route::get('/users/register', [UserController::class, 'register'])->middleware('guest');
+
+// store new user
+Route::post('/users', [UserController::class, 'store']);
+
+// logout
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+// show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 /* 
 // Route Model binding - same as above
